Prevent oscillations of the charge percentage
Due to noise in the battery voltage measurement (probably) the state
of charge may oscillate between two values such as: 93%, then 91%, then
93% again although the battery is strictly discharging. This is purely
a cosmetic issue but it can be addressed quite easily since we know
when the battery is charging or not. Add a BATTERY_PercentMonotonic()
function to take care of that. This costs 28 bytes in the binary size.

Signed-off-by: Jerome Forissier <jerome@forissier.org>